refactor(frontend): split chat-api into lib/chat module and centralise tests
Split lib/chat-api.ts into lib/chat/{types,client,index}.ts so types and
request functions live separately, with a barrel re-export keeping the
public import surface at @/lib/chat.

Move all lib/*.test.ts files to lib/tests/ and switch their imports to
absolute @/lib/... paths.
