+++ frontend/lib/chat/index.ts
@@ -1,0 +1,2 @@
+export * from "@/lib/chat/types";
+export * from "@/lib/chat/client";
